Fix versioning and Load and API
