+++ requirements.txt
@@ -1,0 +1,13 @@
+langchain==1.2.12
+langchain-core==1.2.20
+langchain-openai==1.1.11
+langchain-community==0.4.1
+langchain-qdrant==1.1.0
+langchain-redis
+python-dotenv
+langfuse==3.14.5
+qdrant-client==1.17.1
+requests
+nemoguardrails>=0.20.0,<1.0.0
+backoff
+pyyaml
