--- starter_code/requirements.txt
@@ -1,12 +1,0 @@
-langchain-core
-langchain-openai
-langchain-community
-langchain_qdrant
-langchain_redis
-python-dotenv
-langfuse==2.60.8
-qdrant-client
-requests
-nemoguardrails
-backoff
-pyyaml
